requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
arrow==0.14.7
    # via ics
asgiref==3.3.4
    # via django
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via django-bootstrap4
bleach==3.3.0
    # via wiki
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.6
    # via requests
click==7.1.2
    # via pip-tools
decorator==5.0.7
    # via ipython
django==3.2.10
    # via
    #   -r requirements.in
    #   django-bootstrap4
    #   django--ckeditor
    #   django-classy-tags
    #   django-extensions
    #   django-filter
    #   django-jsonfield
    #   django-mptt
    #   django-nyt
    #   django-picklefield
    #   django-sekizai
    #   djangorestframework
    #   wiki
django-activity-stream==0.10.0
    # via -r requirements.in
django-bootstrap4==3.0.1
    # via -r requirements.in
django-ckeditor==6.4.0
    # via -r requirements.in
django-classy-tags==2.0.0
    # via django-sekizai
django-constance[database]==2.8.0
    # via -r requirements.in
django-crispy-forms==1.13.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-jsonfield==1.4.1
    # via -r requirements.in
django-jsonfield-compat==0.4.4
    # via -r requirements.in
django-mptt==0.11.0
    # via
    #   -r requirements.in
    #   wiki
django-nyt==1.2
    # via wiki
django-picklefield==3.0.1
    # via django-constance
django-resized==0.3.11
    # via -r requirements.in
django-sekizai==2.0.0
    # via wiki
django-simple-history==3.0.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
ics==0.7
    # via -r requirements.in
idna==2.10
    # via requests
influxdb==5.3.1
    # via -r requirements.in
ipython==7.28.0
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.18.0
    # via ipython
markdown==3.2.2
    # via
    #   -r requirements.in
    #   wiki
matplotlib-inline==0.1.2
    # via ipython
msgpack==1.0.2
    # via influxdb
mypy-extensions==0.4.3
    # via -r requirements.in
packaging==20.9
    # via bleach
parso==0.8.2
    # via jedi
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.2
    # via
    #   -r requirements.in
    #   wiki
pip-tools==6.3.0
    # via -r requirements.in
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pygments==2.9.0
    # via ipython
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via
    #   arrow
    #   ics
    #   influxdb
pytz==2021.1
    # via
    #   django
    #   influxdb
redis==3.5.3
    # via -r requirements.in
requests==2.26.0
    # via
    #   -r requirements.in
    #   influxdb
sentry-sdk==1.4.3
    # via -r requirements.in
six==1.15.0
    # via
    #   bleach
    #   django-jsonfield
    #   ics
    #   influxdb
    #   python-dateutil
sorl-thumbnail==12.7.0
    # via wiki
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tatsu==5.6.1
    # via ics
toml==0.10.2
    # via pep517
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==3.10.0.2
    # via -r requirements.in
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wheel==0.37.0
    # via pip-tools
wiki==0.7.7
    # via -r requirements.in

django_email_verification

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools

py~=1.10.0
pip~=22.1
attrs~=20.3.0
pytest~=6.2.5
isort~=5.9.3
setuptools~=60.2.0
flake8~=4.0.1
pyflakes~=2.4.0
pylint~=2.11.1
astroid~=2.8.0
platformdirs~=2.3.0
mccabe~=0.6.1
pluggy~=0.13.1
iniconfig~=1.1.1
wrapt~=1.12.1
PyJWT~=2.4.0

jwt